feat: initial setup with test api endpoint
