--- ..http
@@ -1,11 +1,0 @@
-@__HostAddress = http://localhost:5278
-
-GET {{__HostAddress}}/weather/
-Accept: application/json
-
-###
-
-GET {{__HostAddress}}/weather/philadelphia
-Accept: application/json
-
-###
