--- README.md
@@ -1,2 +1,0 @@
-# do-csharper
-Playground for C# experiments
