--- controllers/WeatherController.cs
@@ -1,37 +1,0 @@
-using Microsoft.AspNetCore.Mvc;
-using _.DTOs;
-
-namespace _.Controllers;
-
-[ApiController]
-[Route("[controller]")]
-public class WeatherController : ControllerBase
-{
-    private static readonly string[] Summaries = new[]
-    {
-        "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
-    };
-
-    [HttpGet(Name = "GetWeatherForecast")]
-    public IEnumerable<WeatherForecast> Get()
-    {
-        var forecast = Enumerable.Range(1, 5).Select(index =>
-            new WeatherForecast
-            (
-                DateOnly.FromDateTime(DateTime.Now.AddDays(index)),
-                Random.Shared.Next(-20, 55),
-                Summaries[Random.Shared.Next(Summaries.Length)]
-            ))
-            .ToArray();
-        
-        return forecast;
-    }
-
-    [HttpGet("philadelphia", Name = "Philadelphia")]
-    public String GetPhiladelphia()
-    {
-        return "It's always sunny in Philadelphia!";
-    }
-
-
-}
