+++ do-csharper.sln
@@ -1,0 +1,24 @@
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.5.2.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "perspectize-be", "perspectize-be\perspectize-be.csproj", "{E740EE3A-0905-C84D-027E-934A57662664}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{E740EE3A-0905-C84D-027E-934A57662664}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E740EE3A-0905-C84D-027E-934A57662664}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E740EE3A-0905-C84D-027E-934A57662664}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E740EE3A-0905-C84D-027E-934A57662664}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {0E587823-4D37-4C0F-B537-8E69C5E6C75B}
+	EndGlobalSection
+EndGlobal
