--- dtos/WeatherForecast.cs
@@ -1,6 +1,0 @@
-namespace _.DTOs;
-
-public record WeatherForecast(DateOnly Date, int TemperatureC, string? Summary)
-{
-    public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
-}
