+++ perspectize-be/DTOs/VideoRequest.cs
@@ -1,0 +1,7 @@
+namespace App.DTOs
+{
+    public class VideoRequest
+    {
+        public string VideoId { get; set; }
+    }
+}
